+++ src/lib/algorithm/cipher/mode/BlockCipherMode.ts
@@ -3,7 +3,7 @@ import type {Word32Array} from "../../../Word32Array";
 
 export interface BlockCipherModeProps {
   cipher: BlockCipher;
-  iv: number[]|undefined;
+  iv: Word32Array|undefined;
 }
 
 /**
@@ -13,7 +13,7 @@ export interface BlockCipherModeProps {
 export class BlockCipherMode {
   protected _props: BlockCipherModeProps;
   protected _cipher: BlockCipher;
-  protected _iv?: number[];
+  protected _iv?: Word32Array;
   
   public constructor(props: BlockCipherModeProps) {
     this._props = props;
@@ -26,13 +26,6 @@ export class BlockCipherMode {
    */
   public processBlock(words: number[], offset: number) {
     return;
-  }
-  
-  /**
-   * @abstract
-   */
-  public generateAuthTag(cipherText: Word32Array): Word32Array|undefined {
-    return undefined;
   }
   
   /**